requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiomysql==0.2.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
blackscholes==0.1.8
ccxt==4.4.21
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
numpy==2.0.2
pandas==2.2.3
propcache==0.2.0
py_lets_be_rational==1.0.1
py_vollib==1.0.1
pycares==4.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.13.1
simplejson==3.19.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.15.5
